pyproject.toml
[tool.poetry]
name = "get-papers"
version = "0.1.0"
description = "CLI to fetch PubMed papers with pharmaceutical affiliations"
authors = ["Your Name <your.email@example.com>"]
packages = [{ include = "get_papers" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.81"
typer = "^0.9.0"
pandas = "^2.2.2"

[tool.poetry.scripts]
get-papers-list = "get_papers.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"